libtraceevent: add option to disable documentation
On some Linux environment builder(i.e. Buildroot) host asciidoc is not
provided since by default all man/docs are not installed to target but
meson.build at the moment build ascii and html documentation and
requires asciidoc making the building to fail. So let's add doc option
set to true by default to let the user to override it and not produce
the documentation.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>